+++ internal/domain/models/product_model.go
