+++ internal/domain/ports/service/product_service.go
@@ -1,7 +1,7 @@
 package service
 
 import (
-	"github.com/D-Sorrow/meli-frescos/internal/domain/product/models"
+	"github.com/D-Sorrow/meli-frescos/internal/domain/models"
 )
 
 type ProductService interface {
