--- internal/domain/product/models/seller.go
@@ -1,9 +1,0 @@
-package models
-
-type Seller struct {
-	Id          int
-	Cid         int
-	CompanyName string
-	Address     string
-	Telephone   string
-}
