--- internal/domain/product/ports/repository/seller_repository.go
@@ -1,9 +1,0 @@
-package repository
-
-import (
-	"github.com/D-Sorrow/meli-frescos/internal/domain/product/models"
-)
-
-type SellerRepository interface {
-	GetSellers() (map[int]models.Seller, error)
-}
