--- internal/domain/product/ports/service/seller_service.go
@@ -1,9 +1,0 @@
-package service
-
-import (
-	"github.com/D-Sorrow/meli-frescos/internal/domain/product/models"
-)
-
-type SellerService interface {
-	GetSellers() (map[int]models.Seller, error)
-}
