+++ internal/infrastructure/loader/mappers/product_mapper.go
@@ -1,7 +1,7 @@
 package mappers
 
 import (
-	"github.com/D-Sorrow/meli-frescos/internal/domain/product/models"
+	"github.com/D-Sorrow/meli-frescos/internal/domain/models"
 	"github.com/D-Sorrow/meli-frescos/internal/infrastructure/loader/entity"
 )
