+++ internal/infrastructure/repository/error_management/product_repository_error.go
