--- internal/transport/handlers/product/dto/seller_dto.go
@@ -1,9 +1,0 @@
-package dto
-
-type SellerDto struct {
-	Id          int
-	Cid         int
-	CompanyName string
-	Address     string
-	Telephone   string
-}
