+++ internal/transport/handlers/product_handler.go
@@ -3,10 +3,10 @@ package product
 import (
 	"encoding/json"
 	"errors"
-	"github.com/D-Sorrow/meli-frescos/internal/domain/product/ports/service"
-	"github.com/D-Sorrow/meli-frescos/internal/infrastructure/product/repository/error_management"
-	dto "github.com/D-Sorrow/meli-frescos/internal/transport/handlers/product/dto"
-	mapper "github.com/D-Sorrow/meli-frescos/internal/transport/handlers/product/mappers"
+	"github.com/D-Sorrow/meli-frescos/internal/domain/ports/service"
+	"github.com/D-Sorrow/meli-frescos/internal/infrastructure/repository/error_management"
+	dto "github.com/D-Sorrow/meli-frescos/internal/transport/handlers/dto"
+	mapper "github.com/D-Sorrow/meli-frescos/internal/transport/handlers/mappers"
 	"github.com/bootcamp-go/web/response"
 	"github.com/go-chi/chi/v5"
 	"net/http"
