--- internal/transport/router/seller_router.go
@@ -1,24 +1,0 @@
-package router
-
-import (
-	"github.com/D-Sorrow/meli-frescos/internal/domain/product/models"
-	"github.com/D-Sorrow/meli-frescos/internal/domain/product/service"
-	"github.com/D-Sorrow/meli-frescos/internal/infrastructure/product/repository"
-	"github.com/D-Sorrow/meli-frescos/internal/transport/handlers"
-	"github.com/go-chi/chi/v5"
-)
-
-func InitSellerRouter(rt *chi.Mux) {
-
-	mapSeller := make(map[int]models.Seller)
-	repositoryImp := repository.NewSellerRepository(mapSeller)
-
-	serviceImp := service.NewSellerService(repositoryImp)
-
-	handler := handlers.NewHandlerService(serviceImp)
-
-	rt.Route("/seller", func(rt chi.Router) {
-		rt.Get("/", handler.GetSeller())
-	})
-	return
-}
